refactor: extract language preference logic into shared module
