+++ index.html
@@ -707,21 +707,7 @@
                 }, 1500);
             }, 3000);
 
-            const languageDropdown = document.getElementById("languagedropdown");
-            if (languageDropdown) {
-            const langLinks = languageDropdown.querySelectorAll("a");
-            langLinks.forEach(link => {
-            link.addEventListener("click", function (e) {
-                e.preventDefault(); // prevent default <a> behavior
-                const selectedLang = this.id; // ID corresponds to language code
-                const currentLang = localStorage.getItem("languagePreference");
-                if (currentLang !== selectedLang) {
-                    localStorage.setItem("languagePreference", selectedLang);
-                    location.reload(); // automatically reload
-                }
-            });
-        });
-    }
+            initializeLanguagePreference();
         });
 
     </script>
@@ -982,7 +968,7 @@
 
 </script>
 
-
+    <script src="js/languagePreference.js"></script>
 </body>
 
 </html>
